smaller types

+++ conversion.rs
@@ -1,15 +1,15 @@
-fn c2f(c: i64) -> f64 {
-    let c = c as f64;
+fn c2f(c: i8) -> f32 {
+    let c = c as f32;
     c * 9.0 / 5.0 + 32.0
 }
 
-fn is_mirror(a: i64, b: i64) -> bool {
+fn is_mirror(a: i8, b: i8) -> bool {
     let a = massage(a);
     let b = reverse_string(massage(b));
     a == b
 }
 
-fn massage(n: i64) -> String {
+fn massage(n: i8) -> String {
     if n < 10 {
         return format!("0{}", n);
     } else if n >= 100 {
@@ -23,7 +23,7 @@ fn reverse_string(s: String) -> String {
     s.chars().rev().collect()
 }
 
-fn print_conversion(c: i64, f: i64) {
+fn print_conversion(c: i8, f: i8) {
     println!("{}°C ~= {}°F", c, f);
 }
 
@@ -31,10 +31,10 @@ fn main() {
     let mut c = 4;
     while c < 100 {
         let f = c2f(c);
-        if is_mirror(c, f.ceil() as i64) {
-            print_conversion(c, f.ceil() as i64)
-        } else if is_mirror(c, f.floor() as i64) {
-            print_conversion(c, f.floor() as i64)
+        if is_mirror(c, f.ceil() as i8) {
+            print_conversion(c, f.ceil() as i8)
+        } else if is_mirror(c, f.floor() as i8) {
+            print_conversion(c, f.floor() as i8)
         } else {
             break;
         }
